+++ CHANGELOG.md
@@ -5,7 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.11.1] - TBD
+## [2.11.2] - TBD
+
+### Added
+
+-   Highlight maxes in weapons page
+-   Unit tests for the Home page
+
+## Changed
+
+-   Updated favicon
+-   Updated application name to Haligtree
+
+## [2.11.1] - 2025-03-31
 
 ### Added